requirements.txt
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain-groq